backend/requirements.txt
Django>=4.2
 djangorestframework>=3.14
 djangorestframework-simplejwt>=5.2
 python-decouple>=3.8
 Pillow>=10.0
 django-cors-headers>=4.2
